Refactor to move AppBar into its own folder
