+++ packages/react-ui/src/index.d.ts
@@ -30,7 +30,7 @@ import Button from './components/atoms/Button';
 import PasswordField from './components/atoms/PasswordField';
 import SelectField from './components/atoms/SelectField';
 import UploadField from './components/molecules/UploadField';
-import AppBar from './components/organisms/AppBar';
+import AppBar from './components/organisms/AppBar/AppBar';
 import LabelWithIndicator from './components/atoms/LabelWithIndicator';
 
 export {
